LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/percentage-size-quirks-002-expected.txt

flex with 1 wrapper div inside:

flex with several wrapper divs inside:


FAIL .pct 1 assert_equals:
<div class="pct" data-expected-height="50"></div>
height expected 50 but got 10
FAIL .pct 2 assert_equals:
<div class="pct" data-expected-height="50"></div>
height expected 50 but got 10
